Added Navigation. Fixed items
