+++ src/components/Navigation.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+ 
+ 
+const Navigation = () => {
+    return (<ul>
+      <li>
+        <Link to={"/"}>Gloomhaven</Link>
+      </li>
+      <li>
+        <Link to={"/jotl"}>JOTL</Link>
+      </li>
+    </ul>)
+};
+ 
+export default Navigation;
